Moved probes from package com.hazelcast.simulator.probes.probes to com.hazelcast.simulator.probes. Renamed XML utils class and moved it to own package.
