+++ probes/src/main/java/com/hazelcast/simulator/probes/Result.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.probes.probes;
+package com.hazelcast.simulator.probes;
 
 import org.HdrHistogram.Histogram;
