+++ probes/src/test/java/com/hazelcast/simulator/probes/HistogramTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.simulator.probes.probes;
+package com.hazelcast.simulator.probes;
 
 import org.HdrHistogram.Histogram;
 import org.junit.Test;
