+++ simulator/src/main/java/com/hazelcast/simulator/worker/TestContainer.java
@@ -1,7 +1,7 @@
 package com.hazelcast.simulator.worker;
 
-import com.hazelcast.simulator.probes.probes.Probe;
-import com.hazelcast.simulator.probes.probes.impl.ProbeImpl;
+import com.hazelcast.simulator.probes.Probe;
+import com.hazelcast.simulator.probes.impl.ProbeImpl;
 import com.hazelcast.simulator.test.TestCase;
 import com.hazelcast.simulator.test.TestContext;
 import com.hazelcast.simulator.test.TestPhase;
