+++ simulator/src/main/java/com/hazelcast/simulator/worker/performance/WorkerPerformanceMonitor.java
@@ -1,6 +1,6 @@
 package com.hazelcast.simulator.worker.performance;
 
-import com.hazelcast.simulator.probes.probes.Probe;
+import com.hazelcast.simulator.probes.Probe;
 import com.hazelcast.simulator.protocol.connector.ServerConnector;
 import com.hazelcast.simulator.protocol.core.SimulatorAddress;
 import com.hazelcast.simulator.protocol.operation.PerformanceStateOperation;
