+++ simulator/src/test/java/com/hazelcast/simulator/tests/PerformanceMonitorTest.java
@@ -15,7 +15,7 @@
  */
 package com.hazelcast.simulator.tests;
 
-import com.hazelcast.simulator.probes.probes.Probe;
+import com.hazelcast.simulator.probes.Probe;
 import com.hazelcast.simulator.test.annotations.Run;
 import com.hazelcast.simulator.test.annotations.SimulatorProbe;
