+++ tests/src/main/java/com/hazelcast/simulator/tests/syntheticmap/StringStringSyntheticMapTest.java
@@ -3,7 +3,7 @@ package com.hazelcast.simulator.tests.syntheticmap;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
-import com.hazelcast.simulator.probes.probes.Probe;
+import com.hazelcast.simulator.probes.Probe;
 import com.hazelcast.simulator.test.TestContext;
 import com.hazelcast.simulator.test.TestRunner;
 import com.hazelcast.simulator.test.annotations.RunWithWorker;
